+++ src/main/java/site/scalarstudios/scalarpower/item/ScalarPowerCreativeTabs.java
@@ -33,6 +33,7 @@ public class ScalarPowerCreativeTabs {
 
             /* Machines */
             event.accept(ScalarPowerBlocks.ALLOY_SMELTER.get());
+            event.accept(ScalarPowerBlocks.CHARGER.get());
             event.accept(ScalarPowerBlocks.EXTRACTOR.get());
             event.accept(ScalarPowerBlocks.FREEZER.get());
             event.accept(ScalarPowerBlocks.LIQUIFIER.get());